fix wrong parameter

+++ de.upb.swt.soot.core/src/main/java/de/upb/swt/soot/core/types/PrimitiveType.java
@@ -164,7 +164,7 @@ public abstract class PrimitiveType extends Type {
 
   private static class ShortType extends PrimitiveType {
     private ShortType() {
-      super("byte");
+      super("short");
     }
 
     @Override
